Refactored StructureController
